+++ src/index.ts
@@ -7,6 +7,7 @@ export { ButtonTypes } from "./components/Button/ButtonTypes";
 export { default as Card } from "./components/Card/Card";
 export { default as Checkbox } from "./components/Checkbox/Checkbox";
 export { default as Heading } from "./components/Heading/Heading";
+export { HeadingDisplaySizes } from "./components/Heading/HeadingDisplaySizes";
 export { default as HelperErrorText } from "./components/HelperErrorText/HelperErrorText";
 export { default as Hero } from "./components/Hero/Hero";
 export { HeroTypes } from "./components/Hero/HeroTypes";
@@ -28,6 +29,12 @@ export { LinkTypes } from "./components/Link/LinkTypes";
 export { default as List } from "./components/List/List";
 export { ListTypes } from "./components/List/ListTypes";
 export { default as Modal } from "./components/Modal/Modal";
+export {
+  default as Notification,
+  NotificationContent,
+  NotificationHeading,
+} from "./components/Notification/Notification";
+export { NotificationTypes } from "./components/Notification/NotificationTypes";
 export { default as Pagination } from "./components/Pagination/Pagination";
 export { default as Radio } from "./components/Radio/Radio";
 export { default as SearchBar } from "./components/SearchBar/SearchBar";